LayoutTests/fast/css/remove-shorthand-expected.txt

Test for http://bugs.webkit.org/show_bug.cgi?id=9284 Quirksmode (CSS1): Removing inline border styles is impossible.

Starting with a declaration containing all properties that are constituents of shortcuts, see what is left when a shortcut property is removed. The shortcut’s constituents and only them should be removed.

Removing 'border' leaves 'background-image: url(file:///background.png); background-repeat: repeat-y; background-attachment: scroll; background-position-x: 100%; background-position-y: 100%; -webkit-background-clip: initial; -webkit-background-origin: initial; background-color: red; list-style-type: square; list-style-position: outside; list-style-image: url(file:///marker.png); margin-top: 5%; margin-right: 5%; margin-bottom: 5%; margin-left: 5%; -webkit-margin-top-collapse: collapse; -webkit-margin-bottom-collapse: collapse; -webkit-marquee-direction: ahead; -webkit-marquee-increment: small; -webkit-marquee-repetition: infinite; -webkit-marquee-style: slide; -webkit-marquee-speed: slow; padding-top: 2em; padding-right: 2em; padding-bottom: 2em; padding-left: 2em; -webkit-text-stroke-color: orange; -webkit-text-stroke-width: 1pt; -webkit-column-width: 100px; -webkit-column-count: 3; -webkit-column-rule-width: thick; -webkit-column-rule-style: dashed; -webkit-column-rule-color: silver; -webkit-border-horizontal-spacing: 10px; -webkit-border-vertical-spacing: 20px; '
Removing 'border-top' leaves 'border-right-width: thin; border-bottom-width: thin; border-left-width: thin; border-right-style: dotted; border-bottom-style: dotted; border-left-style: dotted; border-right-color: purple; border-bottom-color: purple; border-left-color: purple; background-image: url(file:///background.png); background-repeat: repeat-y; background-attachment: scroll; background-position-x: 100%; background-position-y: 100%; -webkit-background-clip: initial; -webkit-background-origin: initial; background-color: red; list-style-type: square; list-style-position: outside; list-style-image: url(file:///marker.png); margin-top: 5%; margin-right: 5%; margin-bottom: 5%; margin-left: 5%; -webkit-margin-top-collapse: collapse; -webkit-margin-bottom-collapse: collapse; -webkit-marquee-direction: ahead; -webkit-marquee-increment: small; -webkit-marquee-repetition: infinite; -webkit-marquee-style: slide; -webkit-marquee-speed: slow; padding-top: 2em; padding-right: 2em; padding-bottom: 2em; padding-left: 2em; -webkit-text-stroke-color: orange; -webkit-text-stroke-width: 1pt; -webkit-column-width: 100px; -webkit-column-count: 3; -webkit-column-rule-width: thick; -webkit-column-rule-style: dashed; -webkit-column-rule-color: silver; -webkit-border-horizontal-spacing: 10px; -webkit-border-vertical-spacing: 20px; '
Removing 'border-right' leaves 'border-top-width: thin; border-bottom-width: thin; border-left-width: thin; border-top-style: dotted; border-bottom-style: dotted; border-left-style: dotted; border-top-color: purple; border-bottom-color: purple; border-left-color: purple; background-image: url(file:///background.png); background-repeat: repeat-y; background-attachment: scroll; background-position-x: 100%; background-position-y: 100%; -webkit-background-clip: initial; -webkit-background-origin: initial; background-color: red; list-style-type: square; list-style-position: outside; list-style-image: url(file:///marker.png); margin-top: 5%; margin-right: 5%; margin-bottom: 5%; margin-left: 5%; -webkit-margin-top-collapse: collapse; -webkit-margin-bottom-collapse: collapse; -webkit-marquee-direction: ahead; -webkit-marquee-increment: small; -webkit-marquee-repetition: infinite; -webkit-marquee-style: slide; -webkit-marquee-speed: slow; padding-top: 2em; padding-right: 2em; padding-bottom: 2em; padding-left: 2em; -webkit-text-stroke-color: orange; -webkit-text-stroke-width: 1pt; -webkit-column-width: 100px; -webkit-column-count: 3; -webkit-column-rule-width: thick; -webkit-column-rule-style: dashed; -webkit-column-rule-color: silver; -webkit-border-horizontal-spacing: 10px; -webkit-border-vertical-spacing: 20px; '
Removing 'border-bottom' leaves 'border-top-width: thin; border-right-width: thin; border-left-width: thin; border-top-style: dotted; border-right-style: dotted; border-left-style: dotted; border-top-color: purple; border-right-color: purple; border-left-color: purple; background-image: url(file:///background.png); background-repeat: repeat-y; background-attachment: scroll; background-position-x: 100%; background-position-y: 100%; -webkit-background-clip: initial; -webkit-background-origin: initial; background-color: red; list-style-type: square; list-style-position: outside; list-style-image: url(file:///marker.png); margin-top: 5%; margin-right: 5%; margin-bottom: 5%; margin-left: 5%; -webkit-margin-top-collapse: collapse; -webkit-margin-bottom-collapse: collapse; -webkit-marquee-direction: ahead; -webkit-marquee-increment: small; -webkit-marquee-repetition: infinite; -webkit-marquee-style: slide; -webkit-marquee-speed: slow; padding-top: 2em; padding-right: 2em; padding-bottom: 2em; padding-left: 2em; -webkit-text-stroke-color: orange; -webkit-text-stroke-width: 1pt; -webkit-column-width: 100px; -webkit-column-count: 3; -webkit-column-rule-width: thick; -webkit-column-rule-style: dashed; -webkit-column-rule-color: silver; -webkit-border-horizontal-spacing: 10px; -webkit-border-vertical-spacing: 20px; '
Removing 'border-left' leaves 'border-top-width: thin; border-right-width: thin; border-bottom-width: thin; border-top-style: dotted; border-right-style: dotted; border-bottom-style: dotted; border-top-color: purple; border-right-color: purple; border-bottom-color: purple; background-image: url(file:///background.png); background-repeat: repeat-y; background-attachment: scroll; background-position-x: 100%; background-position-y: 100%; -webkit-background-clip: initial; -webkit-background-origin: initial; background-color: red; list-style-type: square; list-style-position: outside; list-style-image: url(file:///marker.png); margin-top: 5%; margin-right: 5%; margin-bottom: 5%; margin-left: 5%; -webkit-margin-top-collapse: collapse; -webkit-margin-bottom-collapse: collapse; -webkit-marquee-direction: ahead; -webkit-marquee-increment: small; -webkit-marquee-repetition: infinite; -webkit-marquee-style: slide; -webkit-marquee-speed: slow; padding-top: 2em; padding-right: 2em; padding-bottom: 2em; padding-left: 2em; -webkit-text-stroke-color: orange; -webkit-text-stroke-width: 1pt; -webkit-column-width: 100px; -webkit-column-count: 3; -webkit-column-rule-width: thick; -webkit-column-rule-style: dashed; -webkit-column-rule-color: silver; -webkit-border-horizontal-spacing: 10px; -webkit-border-vertical-spacing: 20px; '
Removing 'border-width' leaves 'border-top-style: dotted; border-right-style: dotted; border-bottom-style: dotted; border-left-style: dotted; border-top-color: purple; border-right-color: purple; border-bottom-color: purple; border-left-color: purple; background-image: url(file:///background.png); background-repeat: repeat-y; background-attachment: scroll; background-position-x: 100%; background-position-y: 100%; -webkit-background-clip: initial; -webkit-background-origin: initial; background-color: red; list-style-type: square; list-style-position: outside; list-style-image: url(file:///marker.png); margin-top: 5%; margin-right: 5%; margin-bottom: 5%; margin-left: 5%; -webkit-margin-top-collapse: collapse; -webkit-margin-bottom-collapse: collapse; -webkit-marquee-direction: ahead; -webkit-marquee-increment: small; -webkit-marquee-repetition: infinite; -webkit-marquee-style: slide; -webkit-marquee-speed: slow; padding-top: 2em; padding-right: 2em; padding-bottom: 2em; padding-left: 2em; -webkit-text-stroke-color: orange; -webkit-text-stroke-width: 1pt; -webkit-column-width: 100px; -webkit-column-count: 3; -webkit-column-rule-width: thick; -webkit-column-rule-style: dashed; -webkit-column-rule-color: silver; -webkit-border-horizontal-spacing: 10px; -webkit-border-vertical-spacing: 20px; '
Removing 'border-style' leaves 'border-top-width: thin; border-right-width: thin; border-bottom-width: thin; border-left-width: thin; border-top-color: purple; border-right-color: purple; border-bottom-color: purple; border-left-color: purple; background-image: url(file:///background.png); background-repeat: repeat-y; background-attachment: scroll; background-position-x: 100%; background-position-y: 100%; -webkit-background-clip: initial; -webkit-background-origin: initial; background-color: red; list-style-type: square; list-style-position: outside; list-style-image: url(file:///marker.png); margin-top: 5%; margin-right: 5%; margin-bottom: 5%; margin-left: 5%; -webkit-margin-top-collapse: collapse; -webkit-margin-bottom-collapse: collapse; -webkit-marquee-direction: ahead; -webkit-marquee-increment: small; -webkit-marquee-repetition: infinite; -webkit-marquee-style: slide; -webkit-marquee-speed: slow; padding-top: 2em; padding-right: 2em; padding-bottom: 2em; padding-left: 2em; -webkit-text-stroke-color: orange; -webkit-text-stroke-width: 1pt; -webkit-column-width: 100px; -webkit-column-count: 3; -webkit-column-rule-width: thick; -webkit-column-rule-style: dashed; -webkit-column-rule-color: silver; -webkit-border-horizontal-spacing: 10px; -webkit-border-vertical-spacing: 20px; '
Removing 'border-color' leaves 'border-top-width: thin; border-right-width: thin; border-bottom-width: thin; border-left-width: thin; border-top-style: dotted; border-right-style: dotted; border-bottom-style: dotted; border-left-style: dotted; background-image: url(file:///background.png); background-repeat: repeat-y; background-attachment: scroll; background-position-x: 100%; background-position-y: 100%; -webkit-background-clip: initial; -webkit-background-origin: initial; background-color: red; list-style-type: square; list-style-position: outside; list-style-image: url(file:///marker.png); margin-top: 5%; margin-right: 5%; margin-bottom: 5%; margin-left: 5%; -webkit-margin-top-collapse: collapse; -webkit-margin-bottom-collapse: collapse; -webkit-marquee-direction: ahead; -webkit-marquee-increment: small; -webkit-marquee-repetition: infinite; -webkit-marquee-style: slide; -webkit-marquee-speed: slow; padding-top: 2em; padding-right: 2em; padding-bottom: 2em; padding-left: 2em; -webkit-text-stroke-color: orange; -webkit-text-stroke-width: 1pt; -webkit-column-width: 100px; -webkit-column-count: 3; -webkit-column-rule-width: thick; -webkit-column-rule-style: dashed; -webkit-column-rule-color: silver; -webkit-border-horizontal-spacing: 10px; -webkit-border-vertical-spacing: 20px; '
Removing 'background' leaves 'border-top-width: thin; border-right-width: thin; border-bottom-width: thin; border-left-width: thin; border-top-style: dotted; border-right-style: dotted; border-bottom-style: dotted; border-left-style: dotted; border-top-color: purple; border-right-color: purple; border-bottom-color: purple; border-left-color: purple; -webkit-background-clip: initial; -webkit-background-origin: initial; list-style-type: square; list-style-position: outside; list-style-image: url(file:///marker.png); margin-top: 5%; margin-right: 5%; margin-bottom: 5%; margin-left: 5%; -webkit-margin-top-collapse: collapse; -webkit-margin-bottom-collapse: collapse; -webkit-marquee-direction: ahead; -webkit-marquee-increment: small; -webkit-marquee-repetition: infinite; -webkit-marquee-style: slide; -webkit-marquee-speed: slow; padding-top: 2em; padding-right: 2em; padding-bottom: 2em; padding-left: 2em; -webkit-text-stroke-color: orange; -webkit-text-stroke-width: 1pt; -webkit-column-width: 100px; -webkit-column-count: 3; -webkit-column-rule-width: thick; -webkit-column-rule-style: dashed; -webkit-column-rule-color: silver; -webkit-border-horizontal-spacing: 10px; -webkit-border-vertical-spacing: 20px; '
Removing 'background-position' leaves 'border-top-width: thin; border-right-width: thin; border-bottom-width: thin; border-left-width: thin; border-top-style: dotted; border-right-style: dotted; border-bottom-style: dotted; border-left-style: dotted; border-top-color: purple; border-right-color: purple; border-bottom-color: purple; border-left-color: purple; background-image: url(file:///background.png); background-repeat: repeat-y; background-attachment: scroll; -webkit-background-clip: initial; -webkit-background-origin: initial; background-color: red; list-style-type: square; list-style-position: outside; list-style-image: url(file:///marker.png); margin-top: 5%; margin-right: 5%; margin-bottom: 5%; margin-left: 5%; -webkit-margin-top-collapse: collapse; -webkit-margin-bottom-collapse: collapse; -webkit-marquee-direction: ahead; -webkit-marquee-increment: small; -webkit-marquee-repetition: infinite; -webkit-marquee-style: slide; -webkit-marquee-speed: slow; padding-top: 2em; padding-right: 2em; padding-bottom: 2em; padding-left: 2em; -webkit-text-stroke-color: orange; -webkit-text-stroke-width: 1pt; -webkit-column-width: 100px; -webkit-column-count: 3; -webkit-column-rule-width: thick; -webkit-column-rule-style: dashed; -webkit-column-rule-color: silver; -webkit-border-horizontal-spacing: 10px; -webkit-border-vertical-spacing: 20px; '
Removing 'list-style' leaves 'border-top-width: thin; border-right-width: thin; border-bottom-width: thin; border-left-width: thin; border-top-style: dotted; border-right-style: dotted; border-bottom-style: dotted; border-left-style: dotted; border-top-color: purple; border-right-color: purple; border-bottom-color: purple; border-left-color: purple; background-image: url(file:///background.png); background-repeat: repeat-y; background-attachment: scroll; background-position-x: 100%; background-position-y: 100%; -webkit-background-clip: initial; -webkit-background-origin: initial; background-color: red; margin-top: 5%; margin-right: 5%; margin-bottom: 5%; margin-left: 5%; -webkit-margin-top-collapse: collapse; -webkit-margin-bottom-collapse: collapse; -webkit-marquee-direction: ahead; -webkit-marquee-increment: small; -webkit-marquee-repetition: infinite; -webkit-marquee-style: slide; -webkit-marquee-speed: slow; padding-top: 2em; padding-right: 2em; padding-bottom: 2em; padding-left: 2em; -webkit-text-stroke-color: orange; -webkit-text-stroke-width: 1pt; -webkit-column-width: 100px; -webkit-column-count: 3; -webkit-column-rule-width: thick; -webkit-column-rule-style: dashed; -webkit-column-rule-color: silver; -webkit-border-horizontal-spacing: 10px; -webkit-border-vertical-spacing: 20px; '
Removing 'margin' leaves 'border-top-width: thin; border-right-width: thin; border-bottom-width: thin; border-left-width: thin; border-top-style: dotted; border-right-style: dotted; border-bottom-style: dotted; border-left-style: dotted; border-top-color: purple; border-right-color: purple; border-bottom-color: purple; border-left-color: purple; background-image: url(file:///background.png); background-repeat: repeat-y; background-attachment: scroll; background-position-x: 100%; background-position-y: 100%; -webkit-background-clip: initial; -webkit-background-origin: initial; background-color: red; list-style-type: square; list-style-position: outside; list-style-image: url(file:///marker.png); -webkit-margin-top-collapse: collapse; -webkit-margin-bottom-collapse: collapse; -webkit-marquee-direction: ahead; -webkit-marquee-increment: small; -webkit-marquee-repetition: infinite; -webkit-marquee-style: slide; -webkit-marquee-speed: slow; padding-top: 2em; padding-right: 2em; padding-bottom: 2em; padding-left: 2em; -webkit-text-stroke-color: orange; -webkit-text-stroke-width: 1pt; -webkit-column-width: 100px; -webkit-column-count: 3; -webkit-column-rule-width: thick; -webkit-column-rule-style: dashed; -webkit-column-rule-color: silver; -webkit-border-horizontal-spacing: 10px; -webkit-border-vertical-spacing: 20px; '
Removing '-webkit-margin-collapse' leaves 'border-top-width: thin; border-right-width: thin; border-bottom-width: thin; border-left-width: thin; border-top-style: dotted; border-right-style: dotted; border-bottom-style: dotted; border-left-style: dotted; border-top-color: purple; border-right-color: purple; border-bottom-color: purple; border-left-color: purple; background-image: url(file:///background.png); background-repeat: repeat-y; background-attachment: scroll; background-position-x: 100%; background-position-y: 100%; -webkit-background-clip: initial; -webkit-background-origin: initial; background-color: red; list-style-type: square; list-style-position: outside; list-style-image: url(file:///marker.png); margin-top: 5%; margin-right: 5%; margin-bottom: 5%; margin-left: 5%; -webkit-marquee-direction: ahead; -webkit-marquee-increment: small; -webkit-marquee-repetition: infinite; -webkit-marquee-style: slide; -webkit-marquee-speed: slow; padding-top: 2em; padding-right: 2em; padding-bottom: 2em; padding-left: 2em; -webkit-text-stroke-color: orange; -webkit-text-stroke-width: 1pt; -webkit-column-width: 100px; -webkit-column-count: 3; -webkit-column-rule-width: thick; -webkit-column-rule-style: dashed; -webkit-column-rule-color: silver; -webkit-border-horizontal-spacing: 10px; -webkit-border-vertical-spacing: 20px; '
Removing '-webkit-marquee' leaves 'border-top-width: thin; border-right-width: thin; border-bottom-width: thin; border-left-width: thin; border-top-style: dotted; border-right-style: dotted; border-bottom-style: dotted; border-left-style: dotted; border-top-color: purple; border-right-color: purple; border-bottom-color: purple; border-left-color: purple; background-image: url(file:///background.png); background-repeat: repeat-y; background-attachment: scroll; background-position-x: 100%; background-position-y: 100%; -webkit-background-clip: initial; -webkit-background-origin: initial; background-color: red; list-style-type: square; list-style-position: outside; list-style-image: url(file:///marker.png); margin-top: 5%; margin-right: 5%; margin-bottom: 5%; margin-left: 5%; -webkit-margin-top-collapse: collapse; -webkit-margin-bottom-collapse: collapse; padding-top: 2em; padding-right: 2em; padding-bottom: 2em; padding-left: 2em; -webkit-text-stroke-color: orange; -webkit-text-stroke-width: 1pt; -webkit-column-width: 100px; -webkit-column-count: 3; -webkit-column-rule-width: thick; -webkit-column-rule-style: dashed; -webkit-column-rule-color: silver; -webkit-border-horizontal-spacing: 10px; -webkit-border-vertical-spacing: 20px; '
Removing 'padding' leaves 'border-top-width: thin; border-right-width: thin; border-bottom-width: thin; border-left-width: thin; border-top-style: dotted; border-right-style: dotted; border-bottom-style: dotted; border-left-style: dotted; border-top-color: purple; border-right-color: purple; border-bottom-color: purple; border-left-color: purple; background-image: url(file:///background.png); background-repeat: repeat-y; background-attachment: scroll; background-position-x: 100%; background-position-y: 100%; -webkit-background-clip: initial; -webkit-background-origin: initial; background-color: red; list-style-type: square; list-style-position: outside; list-style-image: url(file:///marker.png); margin-top: 5%; margin-right: 5%; margin-bottom: 5%; margin-left: 5%; -webkit-margin-top-collapse: collapse; -webkit-margin-bottom-collapse: collapse; -webkit-marquee-direction: ahead; -webkit-marquee-increment: small; -webkit-marquee-repetition: infinite; -webkit-marquee-style: slide; -webkit-marquee-speed: slow; -webkit-text-stroke-color: orange; -webkit-text-stroke-width: 1pt; -webkit-column-width: 100px; -webkit-column-count: 3; -webkit-column-rule-width: thick; -webkit-column-rule-style: dashed; -webkit-column-rule-color: silver; -webkit-border-horizontal-spacing: 10px; -webkit-border-vertical-spacing: 20px; '
Removing '-webkit-text-stroke' leaves 'border-top-width: thin; border-right-width: thin; border-bottom-width: thin; border-left-width: thin; border-top-style: dotted; border-right-style: dotted; border-bottom-style: dotted; border-left-style: dotted; border-top-color: purple; border-right-color: purple; border-bottom-color: purple; border-left-color: purple; background-image: url(file:///background.png); background-repeat: repeat-y; background-attachment: scroll; background-position-x: 100%; background-position-y: 100%; -webkit-background-clip: initial; -webkit-background-origin: initial; background-color: red; list-style-type: square; list-style-position: outside; list-style-image: url(file:///marker.png); margin-top: 5%; margin-right: 5%; margin-bottom: 5%; margin-left: 5%; -webkit-margin-top-collapse: collapse; -webkit-margin-bottom-collapse: collapse; -webkit-marquee-direction: ahead; -webkit-marquee-increment: small; -webkit-marquee-repetition: infinite; -webkit-marquee-style: slide; -webkit-marquee-speed: slow; padding-top: 2em; padding-right: 2em; padding-bottom: 2em; padding-left: 2em; -webkit-column-width: 100px; -webkit-column-count: 3; -webkit-column-rule-width: thick; -webkit-column-rule-style: dashed; -webkit-column-rule-color: silver; -webkit-border-horizontal-spacing: 10px; -webkit-border-vertical-spacing: 20px; '
Removing '-webkit-columns' leaves 'border-top-width: thin; border-right-width: thin; border-bottom-width: thin; border-left-width: thin; border-top-style: dotted; border-right-style: dotted; border-bottom-style: dotted; border-left-style: dotted; border-top-color: purple; border-right-color: purple; border-bottom-color: purple; border-left-color: purple; background-image: url(file:///background.png); background-repeat: repeat-y; background-attachment: scroll; background-position-x: 100%; background-position-y: 100%; -webkit-background-clip: initial; -webkit-background-origin: initial; background-color: red; list-style-type: square; list-style-position: outside; list-style-image: url(file:///marker.png); margin-top: 5%; margin-right: 5%; margin-bottom: 5%; margin-left: 5%; -webkit-margin-top-collapse: collapse; -webkit-margin-bottom-collapse: collapse; -webkit-marquee-direction: ahead; -webkit-marquee-increment: small; -webkit-marquee-repetition: infinite; -webkit-marquee-style: slide; -webkit-marquee-speed: slow; padding-top: 2em; padding-right: 2em; padding-bottom: 2em; padding-left: 2em; -webkit-text-stroke-color: orange; -webkit-text-stroke-width: 1pt; -webkit-column-rule-width: thick; -webkit-column-rule-style: dashed; -webkit-column-rule-color: silver; -webkit-border-horizontal-spacing: 10px; -webkit-border-vertical-spacing: 20px; '
Removing '-webkit-column-rule' leaves 'border-top-width: thin; border-right-width: thin; border-bottom-width: thin; border-left-width: thin; border-top-style: dotted; border-right-style: dotted; border-bottom-style: dotted; border-left-style: dotted; border-top-color: purple; border-right-color: purple; border-bottom-color: purple; border-left-color: purple; background-image: url(file:///background.png); background-repeat: repeat-y; background-attachment: scroll; background-position-x: 100%; background-position-y: 100%; -webkit-background-clip: initial; -webkit-background-origin: initial; background-color: red; list-style-type: square; list-style-position: outside; list-style-image: url(file:///marker.png); margin-top: 5%; margin-right: 5%; margin-bottom: 5%; margin-left: 5%; -webkit-margin-top-collapse: collapse; -webkit-margin-bottom-collapse: collapse; -webkit-marquee-direction: ahead; -webkit-marquee-increment: small; -webkit-marquee-repetition: infinite; -webkit-marquee-style: slide; -webkit-marquee-speed: slow; padding-top: 2em; padding-right: 2em; padding-bottom: 2em; padding-left: 2em; -webkit-text-stroke-color: orange; -webkit-text-stroke-width: 1pt; -webkit-column-width: 100px; -webkit-column-count: 3; -webkit-border-horizontal-spacing: 10px; -webkit-border-vertical-spacing: 20px; '
Removing 'border-spacing' leaves 'border-top-width: thin; border-right-width: thin; border-bottom-width: thin; border-left-width: thin; border-top-style: dotted; border-right-style: dotted; border-bottom-style: dotted; border-left-style: dotted; border-top-color: purple; border-right-color: purple; border-bottom-color: purple; border-left-color: purple; background-image: url(file:///background.png); background-repeat: repeat-y; background-attachment: scroll; background-position-x: 100%; background-position-y: 100%; -webkit-background-clip: initial; -webkit-background-origin: initial; background-color: red; list-style-type: square; list-style-position: outside; list-style-image: url(file:///marker.png); margin-top: 5%; margin-right: 5%; margin-bottom: 5%; margin-left: 5%; -webkit-margin-top-collapse: collapse; -webkit-margin-bottom-collapse: collapse; -webkit-marquee-direction: ahead; -webkit-marquee-increment: small; -webkit-marquee-repetition: infinite; -webkit-marquee-style: slide; -webkit-marquee-speed: slow; padding-top: 2em; padding-right: 2em; padding-bottom: 2em; padding-left: 2em; -webkit-text-stroke-color: orange; -webkit-text-stroke-width: 1pt; -webkit-column-width: 100px; -webkit-column-count: 3; -webkit-column-rule-width: thick; -webkit-column-rule-style: dashed; -webkit-column-rule-color: silver; '
